fix #484; Rolling mean
The rolling mean did a naive rolling_sum / window_size.

The window size however is also dependent on the rolling operation.
It should be decremented per Null value and by the effect at the
edges of the array.

+++ polars/polars-core/src/chunked_array/ops/mod.rs
@@ -72,7 +72,7 @@ pub trait ChunkWindow {
     ///                     `false` -> Any Null in the window leads to a Null in the aggregation result.
     fn rolling_sum(
         &self,
-        _window_size: usize,
+        _window_size: u32,
         _weight: Option<&[f64]>,
         _ignore_null: bool,
         _min_periods: u32,
@@ -100,7 +100,7 @@ pub trait ChunkWindow {
     /// * `min_periods` -  Amount of elements in the window that should be filled before computing a result.
     fn rolling_mean(
         &self,
-        _window_size: usize,
+        _window_size: u32,
         _weight: Option<&[f64]>,
         _ignore_null: bool,
         _min_periods: u32,
@@ -129,7 +129,7 @@ pub trait ChunkWindow {
     /// * `min_periods` -  Amount of elements in the window that should be filled before computing a result.
     fn rolling_min(
         &self,
-        _window_size: usize,
+        _window_size: u32,
         _weight: Option<&[f64]>,
         _ignore_null: bool,
         _min_periods: u32,
@@ -158,7 +158,7 @@ pub trait ChunkWindow {
     /// * `min_periods` -  Amount of elements in the window that should be filled before computing a result.
     fn rolling_max(
         &self,
-        _window_size: usize,
+        _window_size: u32,
         _weight: Option<&[f64]>,
         _ignore_null: bool,
         _min_periods: u32,
@@ -191,7 +191,7 @@ pub trait ChunkWindowCustom<T> {
     /// * `min_periods` -  Amount of elements in the window that should be filled before computing a result.
     fn rolling_custom<F>(
         &self,
-        _window_size: usize,
+        _window_size: u32,
         _weight: Option<&[f64]>,
         _fold_fn: F,
         _init_fold: InitFold,
